mads

+++ A5/include/binary_search.h
@@ -1,6 +1,6 @@
 #ifndef BINARY_SEARCH_H
 #define BINARY_SEARCH_H
 
-int binary_search(double [], size_t, size_t, double);
+int binary_search(double [], unsigned long, unsigned long, double);
 
 #endif
