+++ A5/include/bubble_sort.h
@@ -2,6 +2,6 @@
 #ifndef BUBBLE_SORT_H
 #define BUBBLE_SORT_H
 
-void bubble_sort(double [], size_t);
+void bubble_sort(double [], unsigned long);
 
 #endif
