+++ A5/include/get_low.h
@@ -1,6 +1,6 @@
 #ifndef GET_LOW_H
 #define GET_LOW_H
 
-double get_low(double [], size_t);
+double get_low(double [], unsigned long);
 
 #endif
