+++ A5/include/pop_arr.h
@@ -2,6 +2,6 @@
 #ifndef POP_ARR_H
 #define POP_ARR_H
 
-void pop_arr(double [], size_t );
+void pop_arr(double [], unsigned long);
 
 #endif
