+++ A5/include/swap_elm.h
@@ -2,6 +2,6 @@
 #define SWAP_ELM_H
 
 
-void swap_elm(double [], size_t , size_t);
+void swap_elm(double [], unsigned long, unsigned long);
 
 #endif
